feat: add audio handling and synchronization services
- Implemented AudioHandler for managing audio playback, including prayer functionalities and background looping.
- Added SyncService for handling data synchronization with the server, including version checks, media downloads, and data deletion.
- Updated pubspec.lock with new package versions for dependencies.
